requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
slowapi
python-dotenv
